[refactor] - 토큰 대기열 기능 삭제 - 대기열 기능 변경에 따른 예약 도메인 의존성 변경
