+++ src/main/java/kr/hhplus/be/server/api/reservation/application/service/ReservationService.java
@@ -1,6 +1,7 @@
-package kr.hhplus.be.server.api.reservation.domain.service;
+package kr.hhplus.be.server.api.reservation.application.service;
 
 import kr.hhplus.be.server.api.reservation.application.dto.ReservationDto;
+import kr.hhplus.be.server.api.reservation.application.dto.ReservationResult;
 import kr.hhplus.be.server.api.reservation.domain.entity.Reservation;
 import kr.hhplus.be.server.api.reservation.domain.repository.ReservationRepository;
 import kr.hhplus.be.server.common.exception.CustomException;
@@ -19,7 +20,7 @@ public class ReservationService {
 
 	private final ReservationRepository reservationRepository;
 
-	public Reservation makeReservation(ReservationDto dto) {
+	public ReservationResult makeReservation(ReservationDto dto) {
 		Reservation reservation = Reservation.builder().
 				userId(dto.userId()).
 				seatId(dto.seatId()).
@@ -29,18 +30,16 @@ public class ReservationService {
 				isReserved(true).
 				expiredAt(LocalDateTime.now().plusMinutes(5)).
 				build();
-        reservationRepository.save(reservation);
-		return reservation;
+        return ReservationResult.from(reservationRepository.save(reservation));
 	}
 
-	public Reservation confirmReservation(long id) {
+	public ReservationResult confirmReservation(long id) {
 		Reservation reservation = reservationRepository.findById(id);
 		if (reservation.getExpiredAt().isBefore(LocalDateTime.now())) {
 			throw new CustomException(ErrorCode.EXPIRED_SEAT);
 		}
 		reservation.confirm();
-		reservationRepository.save(reservation);
-		return reservation;
+		return ReservationResult.from(reservationRepository.save(reservation));
 	}
 
 	public List<Reservation> getExpiredReservations() {
